ExtractExtensions is a TransformerFactory which extracts a set of
extensions from transformed mesages. An instance of this transformer
should only be used to process a single message, after which the
extracted extensions can be read from the map. All extension which are
present as keys in the map will be extracted and other extensions will
be ignored.

Signed-off-by: Ian Milligan <ianmllgn@gmail.com>