@ian-mi 2020-04-07
This should be superior in the case that you need to gather multiple extensions (which I will need to for tracing). Otherwise you would need to use multiple factories adding additional nested calls to every writer call as well as performing multiple string comparisons on every SetExtension.