@slinkydeveloper 2020-04-07
> Otherwise you would need to use multiple factories adding additional nested calls to every writer call as well as performing multiple string comparisons on every SetExtension.

Yeah my guess is that string comparisons are cheaper, but i wanna try it before reaching a conclusion