@slinkydeveloper 2020-04-07
So I've tried to create a `ExtractExtension` of my own with single extension extraction. old is this patch and new is my patch:

```
benchmark                                   old ns/op     new ns/op     delta
BenchmarkHttpWithToEvent-8                  21728         20512         -5.60%
BenchmarkNoDataHttpWithToEvent-8            20200         20828         +3.11%
BenchmarkHttpWithBuffering-8                11823         11926         +0.87%
BenchmarkHttpWithDirect-8                   9911          9692          -2.21%
BenchmarkNoDataHttpWithBuffering-8          11737         11645         -0.78%
BenchmarkNoDataHttpWithDirect-8             9728          9668          -0.62%

benchmark                                   old allocs     new allocs     delta
BenchmarkHttpWithToEvent-8                  151            151            +0.00%
BenchmarkNoDataHttpWithToEvent-8            145            145            +0.00%
BenchmarkHttpWithBuffering-8                121            122            +0.83%
BenchmarkHttpWithDirect-8                   114            115            +0.88%
BenchmarkNoDataHttpWithBuffering-8          119            120            +0.84%
BenchmarkNoDataHttpWithDirect-8             112            113            +0.89%

benchmark                                   old bytes     new bytes     delta
BenchmarkHttpWithToEvent-8                  10500         10496         -0.04%
BenchmarkNoDataHttpWithToEvent-8            9864          9864          +0.00%
BenchmarkHttpWithBuffering-8                6086          6109          +0.38%
BenchmarkHttpWithDirect-8                   4760          4791          +0.65%
BenchmarkNoDataHttpWithBuffering-8          6038          6068          +0.50%
BenchmarkNoDataHttpWithDirect-8             4712          4744          +0.68%

```

Interesting enough, having this map allocates less, but in some cases it consumes a little bit more in terms of cpu time (but, honestly to me that delta looks like measurement error).

So yeah, let's go with the map. Please apply the changes to the tests and then we merge it