@slinkydeveloper 2020-04-08
I think this one is invalid now. Because now it's more easy to implement the extraction of extensions with transformers, i close this one in favour of https://github.com/cloudevents/sdk-go/pull/447 which is more targeted to our need